requirements.txt
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.57
botocore==1.34.57
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.27.1
